remove 5.10-ism
Possessive quantifier syntax was introduced in that version as syntactic
sugar for the much bulkier (?>) construct.

All patterns in this module are written to minimise the incidence of
adjacent atoms that can match at the same string position, to minimise
backtracking.

Minimising the remaining backtracking is a microöptimsation worth using
if it does not diminish the readability of the code: since Perl 5.10 it
has been, but prior to that (as this module targets) it was not.